vidya-ai/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
av==12.3.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
comtypes==1.4.16
ctranslate2==4.7.1
dataclasses-json==0.6.7
diskcache==5.6.3
faiss-cpu==1.9.0.post1
fastapi==0.135.3
faster-whisper==1.0.2
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.20.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
llama_cpp_python==0.2.77
loguru==0.7.3
lxml==6.0.3
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.9.4
numpy==1.26.4
onnxruntime==1.24.4
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pillow==12.2.0
propcache==0.4.1
protobuf==7.34.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyMuPDF==1.27.2.2
pypiwin32==223
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.24
python-pptx==1.0.2
pyttsx3==2.90
pywin32==311
PyYAML==6.0.3
rank-bm25==0.2.2
regex==2026.4.4
reportlab==4.4.10
requests==2.33.1
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==2.2.2
sentencepiece==0.2.1
SQLAlchemy==2.0.49
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.15.0
torch==2.11.0+cpu
torchvision==0.26.0
tqdm==4.67.3
transformers==4.36.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
win32_setctime==1.2.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
